+++ pkg/subDomainScan/Result.go
@@ -3,4 +3,5 @@ package subDomainScan
 type Result struct {
 	subdomainsFromTlsScan map[string]struct{}
 	subdomainsFromCrtSh   map[string]struct{}
+	crtShStatus           crtShStatus
 }